File/class renaming around chat rooms
